+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Added
 - new `sc4pac reinstall` command to force re-installing individual packages, e.g. to restore missing files
+- new `sc4pac repair` command to scan the Plugins folder for .sc4pac folders that are out-of-sync with sc4pac's internal state
 
 ### Fixed
 - an issue preventing the import and export of JSON mod sets if the JSON file was too large ([#39][gui39])